+++ CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add simple REST server implementation
 - Add reusable snippets that are reused across the docs
 - Added snippet for connecting to superduperdb in docs
+- Added support to serialize documents in a flat way "_leaves"
 
 #### Bug Fixes
 - Separate nightly release from sandbox